+++ lib/sequent/core/helpers/message_matchers/class_matcher.rb
